Add post-upload delay
When uploading concatenated UF2 files with different family ids, it's important
that the bootloader continue receiving the rest of the file from the host, even
if it knows that there's no more relevant blocks: otherwise the host operating
system will assume an error and show the user a scary warning message.

Wait a little bit after finishing the update process (and continue to process
USB) to give this a chance to happen.